refactor: 폴더 구조 변경
도메인형 폴더 구조로 변경

Related to: #7
