+++ guidely/src/main/java/dotori/guidely/domain/oauth/OAuthApiClient.java
@@ -1,5 +1,8 @@
 package dotori.guidely.domain.oauth;
 
+import dotori.guidely.domain.oauth.domain.OAuthInfoResponse;
+import dotori.guidely.domain.oauth.domain.OAuthProvider;
+
 public interface OAuthApiClient {
     OAuthProvider oAuthProvider();
