+++ guidely/src/main/java/dotori/guidely/domain/oauth/OAuthLoginParams.java
@@ -1,5 +1,6 @@
 package dotori.guidely.domain.oauth;
 
+import dotori.guidely.domain.oauth.domain.OAuthProvider;
 import org.springframework.util.MultiValueMap;
 
 public interface OAuthLoginParams {
