--- guidely/src/main/java/dotori/guidely/domain/oauth/OAuthProvider.java
@@ -1,5 +1,0 @@
-package dotori.guidely.domain.oauth;
-
-public enum OAuthProvider {
-    KAKAO
-}
