+++ guidely/src/main/java/dotori/guidely/domain/oauth/domain/OAuthInfoResponse.java
@@ -1,4 +1,4 @@
-package dotori.guidely.domain.oauth;
+package dotori.guidely.domain.oauth.domain;
 
 public interface OAuthInfoResponse {
     String getEmail();
