+++ guidely/src/main/java/dotori/guidely/domain/oauth/domain/OAuthProvider.java
@@ -1,0 +1,5 @@
+package dotori.guidely.domain.oauth.domain;
+
+public enum OAuthProvider {
+    KAKAO
+}
