+++ guidely/src/main/java/dotori/guidely/global/config/SwaggerConfiguration.java
@@ -1,4 +1,4 @@
-package dotori.guidely.config;
+package dotori.guidely.global.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
